Fixed bugs in the separate round registration
Fixed bugs in find spots which was often causing the spot_no array to be incorrect. Added parameters to the config file in a new section called reg_to_anchor_info to allow the user to specify the patches for comparison of their images. Fixed bugs in the separate round registration that were causing incorrect array sizes due to former versions of the spot_details array.

TO DO: Restructure the sep_round_reg data to make it more like the run section of the pipeline.