Fix to bug in register_initial
This bug has been showing up a lot lately. When running register_initial, users will get to a tile and round and suddenly the code quits with the message below at line ~477 in stitch/shift.

IndexError: too many indices for array: array is 0-dimensional, but 1 were indexed.

I believe this is due to a bug at approximately line 424, since we were previously allowing the program to pad a 'None' variable with a 0 at the end. The result of this is that shift_thresh was being returned as [None], which is an ndarray, and not None, so later checks were asserting that shift_thresh is not None, but when trying to call shift_thresh returning an error.